+++ ds_judgements_public_ui/templates/components/heading.jinja
@@ -1,4 +1,5 @@
-{% macro heading(tag="h1", size="default", id=None, withoutMargins=False) %}
+{% macro heading(tag="h1", size="default", id=None, withoutMargins=False, content=None) %}
+  {% set body_content = caller() if caller is defined else content if content is defined else "" %}
   {% set size_classes = {
     "default": "",
     "h1": "heading heading--h1",
@@ -8,17 +9,25 @@
   } %}
   {% set size_class = size_classes.get(size, size_classes["default"]) %}
   {% set margin_class = "heading--without-margins" if withoutMargins else "" %}
-  {% set text = caller() | trim %}
+  {% set text = body_content | trim %}
   {% set id_value = id | default(text | hyphenate, true) %}
-  {% if tag == "h1" %}<h1 id="{{ id_value }}" class="{{ size_class }} {{ margin_class }}">{{ caller() }}</h1>{% endif %}
-  {% if tag == "h2" %}<h2 id="{{ id_value }}" class="{{ size_class }} {{ margin_class }}">{{ caller() }}</h2>{% endif %}
-  {% if tag == "h3" %}<h3 id="{{ id_value }}" class="{{ size_class }} {{ margin_class }}">{{ caller() }}</h3>{% endif %}
-  {% if tag == "h4" %}<h4 id="{{ id_value }}" class="{{ size_class }} {{ margin_class }}">{{ caller() }}</h4>{% endif %}
+  {% if tag == "h1" %}
+    <h1 id="{{ id_value }}" class="{{ size_class }} {{ margin_class }}">{{ body_content }}</h1>
+  {% endif %}
+  {% if tag == "h2" %}
+    <h2 id="{{ id_value }}" class="{{ size_class }} {{ margin_class }}">{{ body_content }}</h2>
+  {% endif %}
+  {% if tag == "h3" %}
+    <h3 id="{{ id_value }}" class="{{ size_class }} {{ margin_class }}">{{ body_content }}</h3>
+  {% endif %}
+  {% if tag == "h4" %}
+    <h4 id="{{ id_value }}" class="{{ size_class }} {{ margin_class }}">{{ body_content }}</h4>
+  {% endif %}
   {% if tag == "div" %}
-    <div id="{{ id_value }}" class="{{ size_class }} {{ margin_class }}">{{ caller() }}</div>
+    <div id="{{ id_value }}" class="{{ size_class }} {{ margin_class }}">{{ body_content }}</div>
   {% endif %}
   {% if tag == "span" %}
-    <span id="{{ id_value }}" class="{{ size_class }} {{ margin_class }}">{{ caller() }}</span>
+    <span id="{{ id_value }}" class="{{ size_class }} {{ margin_class }}">{{ body_content }}</span>
   {% endif %}
 {% endmacro %}
 {% macro heading_examples() %}
