+++ judgments/views/about_this_service/__init__.py
@@ -1,12 +1,11 @@
 from .base import AboutThisServiceView
-from .courts_and_date_coverage import CourtsAndDateCoverageView, CourtsAndTribunalsInFclView
+from .courts_and_coverage import CourtsAndCoverageView
 from .publishing_policy import PublishingPolicyView
 from .what_we_provide import AboutFindCaseLawView, WhatWeProvideView
 
 __all__ = [
     "AboutThisServiceView",
-    "CourtsAndDateCoverageView",
-    "CourtsAndTribunalsInFclView",
+    "CourtsAndCoverageView",
     "PublishingPolicyView",
     "AboutFindCaseLawView",
     "WhatWeProvideView",
