--- judgments/views/about_this_service/courts_and_date_coverage/__init__.py
@@ -1,4 +1,0 @@
-from .base import CourtsAndDateCoverageView
-from .courts_and_tribunals_in_find_case_law import CourtsAndTribunalsInFclView
-
-__all__ = ["CourtsAndDateCoverageView", "CourtsAndTribunalsInFclView"]
