+++ judgments/views/search_and_browse/__init__.py
@@ -1,5 +1,4 @@
 from .base import SearchAndBrowseView
-from .browse_courts_and_tribunals import BrowseCourtsAndTribunalsView
 from .how_to_search_find_case_law import HowToSearchFindCaseLawView
 
-__all__ = ["SearchAndBrowseView", "BrowseCourtsAndTribunalsView", "HowToSearchFindCaseLawView"]
+__all__ = ["SearchAndBrowseView", "HowToSearchFindCaseLawView"]
